Hack to fix CSP issues with ag-grid (see issue 1093)
As described in [issue 1093](https://github.com/ceolter/ag-grid/issues/1093),
the inline styles in borderLayout violate Content Security Policy rules.

This branch is a hack of the dist files to move the styles to the css
file. It is **not** a complete fix; it only covers the specific case
required, and does so in the `dist` files rather than the `src`. This
should not be used for any other purpose!

The expectation is that this will be fixed properly in the main
repo and we can revert to using that.

Test Plan
- The same fix has been tested in the end project and appears to work well